#4 [MOVE] : example_compose -> example.compose 위치 이동
